requirements.txt
gspread==5.7.2
google-auth==2.23.4
pandas==2.1.1
